Add basic files to project
Add project files for loading javafx and a basic Java class file for verifying that the modules are loaded and can be used.
Future commit will build upon the SketchPractice Java class file, and run it as the main program.

Some things to note are that running solely via command line and adding modules via run configurations, which a lot of tutorials do, doesn't seem to work, at least in IntelliJ