Add Observer and Observable Design pattern, reorganize files into model-view-controller and utility based packages
Added Files that implemented the Observer and Observable Design pattern and began development of app based on the model view controller diagram.
The ImageCyclingModel class will be an abstract (not abstract class) representation of the functionality of the app, and is still incomplete.
More files will be added that call upon information and methods of the Image Cycling Model and represent its functionality in a GUI.
The SketchPractice class has been modified to include a test implementation for selecting directories, and by no means represents the final product.
As such, the SketchPractice class will be vastly different in production.