Refactor code for the controller class to not have to refer to store an instance of view
After careful consideration, it seemed impossible to completely separate the tight coupling of the view and controller
due to how event handlers are written and may need to do more complicated tasks such as refer to another gui component
within the GUI class. As such, most event handler classes from CyclerController have been removed.

The intended usage now is to have the Gui attach itself to the model clases outside of the Cycler Controller classes (though
there is a function for doing so within the CyclerController class after such attachment may or may not have been done).
With this arrangement, at the minimum, the CyclerController does not need to be dependent on a specific GUI class or view
to operate, and can be extended to do more things to the model (or add more models and functions) if needed, without worrying
about modifying a GUI, as that controller logic has been left to the JavaFX event handlers.

If there was a way to make this better, I would gladly implement it, but I do not yet know at this time.