Complete Functional App with unformatted UI
I'm too tired to commit individually and explain each change as it is literally 4AM
Basically, we have finally completed a somewhat functional app that can cycle through the images
selected by the user.

Some issues are:
- code needs cleanup
- need to address remaining todos in comments, such as what to do when an image is deleted while cycling through them, to the point where no image is accessible
- The GUI needs to be a bit more functional when it comes to size and formatting. Currently, stage opens at its default size and nothing is calculated
based on the actual width or height of the monitor. We'll address as much of this as we can with css first. For now, resize the window manually.

There is still much work to be done. Good night.