Move Image opening code to view related files, ensure filestream is closed after usage
This also includes an update to how moved or deleted images are handled. A stackpane is used
with hidden error text that shows only if no image is being rendered.

Not closing the filestream in previous commits was a mistake. If not caught, too many system resources may have been occupied.
As such, always remember to close the filestream after it's loaded as an Image.